+++ src/core/components/forms/checkboxCard/checkboxCard.stories.tsx
@@ -25,6 +25,18 @@ export const Default: Story = {
     args: {
         avatar: 'https://assets-global.website-files.com/5e997428d0f2eb13a90aec8c/63f47db62df04b569e4e004e_icon_aragon.svg',
         label: 'Checkbox label',
+        tag: { label: 'Tag', variant: 'info' },
+    },
+};
+
+/**
+ * Default usage of the CheckboxCard component
+ */
+export const WithDescription: Story = {
+    render: (props) => <CheckboxCard {...props} />,
+    args: {
+        avatar: 'https://assets-global.website-files.com/5e997428d0f2eb13a90aec8c/63f47db62df04b569e4e004e_icon_aragon.svg',
+        label: 'Checkbox label',
         description: 'Checkbox description',
         tag: { label: 'Tag', variant: 'info' },
     },