+++ src/core/components/forms/checkboxGroup/checkboxGroup.stories.tsx
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { Checkbox } from '../checkbox';
 import { CheckboxCard } from '../checkboxCard';
-import { InputNumber } from '../inputNumber';
 import { CheckboxGroup } from './checkboxGroup';
 
 const meta: Meta<typeof CheckboxGroup> = {
@@ -54,40 +53,6 @@ export const CheckboxCardVariant: Story = {
                 avatar="third"
                 tag={{ label: 'First', variant: 'primary' }}
             />
-        </CheckboxGroup>
-    ),
-};
-
-/**
- * Usage of the CheckboxGroup component with CheckboxCard variant and additional content when checked.
- */
-export const CheckboxCardVariantWithChildren: Story = {
-    render: (props) => (
-        <CheckboxGroup {...props}>
-            <CheckboxCard
-                label="First label"
-                description="First description"
-                avatar="first"
-                tag={{ label: 'First', variant: 'info' }}
-            >
-                <InputNumber placeholder="0" />
-            </CheckboxCard>
-            <CheckboxCard
-                label="Second label"
-                description="Longer description for second checkbox"
-                avatar="second"
-                tag={{ label: 'Second', variant: 'warning' }}
-            >
-                <div>Some additional content</div>
-            </CheckboxCard>
-            <CheckboxCard
-                label="Third label"
-                description="First description"
-                avatar="third"
-                tag={{ label: 'First', variant: 'primary' }}
-            >
-                <div>Some additional content</div>
-            </CheckboxCard>
         </CheckboxGroup>
     ),
 };