+++ src/core/components/forms/radioGroup/radioGroup.stories.tsx
@@ -1,6 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { useState } from 'react';
-import { InputNumber } from '../inputNumber';
 import { Radio } from '../radio';
 import { RadioCard } from '../radioCard';
 import { RadioGroup } from './radioGroup';
@@ -64,43 +63,6 @@ export const RadioCardVariant: Story = {
                 avatar="bronze"
                 tag={{ label: 'Bronze', variant: 'warning' }}
             />
-        </RadioGroup>
-    ),
-    args: {
-        defaultValue: '2',
-        name: 'Options',
-        disabled: false,
-        onValueChange: undefined,
-    },
-};
-
-/**
- * Usage of the `RadioGroup` component with the RadioCard and children when selected
- */
-export const RadioCardVariantWithChildren: Story = {
-    render: (props) => (
-        <RadioGroup {...props}>
-            <RadioCard
-                label="Option one"
-                description="The best option"
-                value="1"
-                avatar="gold"
-                tag={{ label: 'Gold', variant: 'success' }}
-            >
-                <InputNumber placeholder="0" />
-            </RadioCard>
-            <RadioCard label="Option two" value="2" avatar="silver" tag={{ label: 'Silver' }}>
-                <div>Some additional content</div>
-            </RadioCard>
-            <RadioCard
-                label="Option three"
-                description="The 3rd best option"
-                value="3"
-                avatar="bronze"
-                tag={{ label: 'Bronze', variant: 'warning' }}
-            >
-                <div>Some additional content</div>
-            </RadioCard>
         </RadioGroup>
     ),
     args: {