+++ CHANGELOG.md
@@ -13,7 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
--   Update core component `RadioCard` to have optional description property and correct alignment when no description
+-   Update core components `RadioCard` and `CheckboxCard` to have optional description property and correct alignment when no description
     provided
 -   Bump `actions/checkout` from 4.2.0 to 4.2.1
 -   Update minor and patch NPM dependencies